+++ src/mdf_viewer/view/widgets/__init__.py
@@ -2,14 +2,14 @@
 
 from mdf_viewer.view.widgets.busy_cursor import busy_cursor
 from mdf_viewer.view.widgets.color_swatch import ColorSwatch
-from mdf_viewer.view.widgets.icons import _icon_suffix, _load_icon
+from mdf_viewer.view.widgets.icons import _icon_color, _load_icon
 from mdf_viewer.view.widgets.splitter import make_splitter
 from mdf_viewer.view.widgets.visibility_toggle_button import VisibilityToggleButton
 
 __all__ = [
     "ColorSwatch",
     "VisibilityToggleButton",
-    "_icon_suffix",
+    "_icon_color",
     "_load_icon",
     "busy_cursor",
     "make_splitter",